Add files via upload
Fun little project that took me an hour to do that, basically, lets you auto swipe right on yubo (if you're using TeamViewer or a similar program with a window of the phone screen). In order to get it to work, you'll need to go into the code and edit the dimensions - use the commented lines of code that will let you accomplish this.

I might, however, come back to this project and let the user input (if they know the dimensions themselves) the dimensions, further removing some of the hassle of opening up the script and editing it in there.